requirements.txt
altgraph==0.17.5
bottle==0.13.4
macholib==1.16.4
modulegraph==0.19.7
packaging==25.0
proxy_tools==0.1.0
py2app==0.28.9
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Security==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-WebKit==12.1
PyQt6==6.10.1
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.3
pywebview==6.1
QtAwesome==1.4.0
QtPy==2.4.3
setuptools==80.9.0
typing_extensions==4.15.0
